+++ src/jsx/components/library/song/Entry.jsx
@@ -10,7 +10,7 @@ import { addSongToPlaylist } from 'actions/player'
 import { clearSongListNotification } from 'actions/library'
 import Song from 'components/song/Song'
 import SongEntryExpanded from './EntryExpanded'
-import { IsPlaylistUser } from 'components/permissions/Playlist'
+import { IsPlaylistUser, KaraStatusIsNotStopped } from 'components/permissions/Playlist'
 import Notification from 'components/generics/Notification'
 import PlayQueueInfo from 'components/song/PlayQueueInfo'
 import DisabledFeedback from 'components/song/DisabledFeedback'
@@ -100,18 +100,20 @@ class SongEntry extends Component {
                             className="controls"
                             id={`song-${this.props.song.id}`}
                         >
-                            <IsPlaylistUser>
-                                <button
-                                    className="control primary"
-                                    onClick={() => {
-                                        this.props.addSongToPlaylist(this.props.song.id)
-                                    }}
-                                >
-                                    <span className="icon">
-                                        <i className="fa fa-plus"></i>
-                                    </span>
-                                </button>
-                            </IsPlaylistUser>
+                            <KaraStatusIsNotStopped>
+                                <IsPlaylistUser>
+                                    <button
+                                        className="control primary"
+                                        onClick={() => {
+                                            this.props.addSongToPlaylist(this.props.song.id)
+                                        }}
+                                    >
+                                        <span className="icon">
+                                            <i className="fa fa-plus"></i>
+                                        </span>
+                                    </button>
+                                </IsPlaylistUser>
+                            </KaraStatusIsNotStopped>
                         </div>
                         <Notification
                             alterationStatus={this.props.addSongStatus}