+++ CHANGELOG.md
@@ -35,6 +35,7 @@
 - Highlight query terms in artists and works libraries.
 - Songs with disabled tags have a visual indicator.
 - Disabled tags are grayed out.
+- Kara status setting
 
 ### Changed